fix(H-06): make root + standalone-workspace strict clippy green
Closes audit H-06 quick-win #2 ('Refactor argument groups; remove
casts/dead function; add standalone lint jobs') and the strict-clippy
findings the audit enumerated:

* kernel_bpf/src/loader/reloc.rs:164,256 — apply_relocations and
  relocate_call each took seven positional usize arguments and tripped
  clippy::too_many_arguments. Collapse the per-call section triple
  (section_idx, section_base, root_section_idx) into a SectionContext
  value, threaded through both methods. Same behavior, smaller surface.
* userspace/rk_bridge/src/ringbuf.rs:114,148 — sys_bpf returns i32, so
  the trailing 'as i32' on -result was a no-op cast that clippy::all
  flagged. Drop the cast.
* userspace/rk_cli/src/config.rs:19 — programs_dir had no callers and
  emitted a dead_code warning. Delete; reintroduce from the command
  path with tests in a follow-up.

Verification (locally, against nightly-2026-07-02 from
rust-toolchain.toml):

* cargo fmt --all -- --check          pass
* cargo clippy -p kernel_bpf     --no-default-features     --features embedded-profile --lib     -- -D clippy::all                 pass
* cargo clippy -p kernel_bpf     --no-default-features     --features cloud-profile --lib     -- -D clippy::all                 pass
* cargo clippy --manifest-path     userspace/rk_bridge/Cargo.toml --lib     -- -D clippy::all                 pass
* cargo clippy --manifest-path     userspace/rk_cli/Cargo.toml --     -D clippy::all                    pass
* cargo test -p kernel_bpf     --no-default-features     --features embedded-profile --lib  390/390
* cargo test -p kernel_bpf     --no-default-features     --features cloud-profile --lib    384/384
* cargo test --manifest-path     userspace/rk_bridge/Cargo.toml    18/18

No production behavior change.

+++ userspace/rk_bridge/src/ringbuf.rs
@@ -111,7 +111,7 @@ impl RingBufConsumer {
                 return Ok(Some(self.event_buf[..size].to_vec()));
             }
 
-            let err = io::Error::from_raw_os_error((-result) as i32);
+            let err = io::Error::from_raw_os_error(-result);
             if err.raw_os_error() == Some(libc::ENOSPC) {
                 let next_len = self.event_buf.len().saturating_mul(2).max(DEFAULT_EVENT_BUF_SIZE);
                 self.event_buf.resize(next_len, 0);
@@ -145,7 +145,7 @@ fn bpf_obj_get(path: &CString) -> io::Result<u32> {
 
     let result = sys_bpf(BPF_OBJ_GET, &attr);
     if result < 0 {
-        Err(io::Error::from_raw_os_error((-result) as i32))
+        Err(io::Error::from_raw_os_error(-result))
     } else {
         Ok(result as u32)
     }
